+++ CHANGES.md
@@ -4,6 +4,7 @@
   - changed `ADD COLUMN DEFAULT NULL` to safe operation for code default
   - changed `ADD COLUMN DEFAULT NOT NULL` to safe operation for `db_default` in django 5.0+
   - added `ZERO_DOWNTIME_MIGRATIONS_KEEP_DEFAULT` settings and changed `ADD COLUMN DEFAULT NOT NULL` with this settings to safe operation for django<5.0
+  - fixed sqlmigrate in idempotent mode
   - updated unsafe migrations links to documentation
   - updated patched code to latest django version
   - improved README