+++ CHANGES.md
@@ -1,6 +1,6 @@
 # django-pg-zero-downtime-migrations changelog
 
-## Unreleased
+## 0.17
   - dropped support for Python 3.6 and 3.7
   - dropped support for Django 3.2, 4.0 and 4.1